+++ tests/test_skoolhtml.py
@@ -4459,7 +4459,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="30000"></span><a href="../asm/30000.html">30000</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">Routine</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30000.html">Routine</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4470,7 +4470,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-b"><span id="30001"></span><a href="../asm/30001.html">30001</a></td>
             <td class="map-length-0">2</td>
             <td class="map-b-desc">
-            <div class="map-entry-title-10">Bytes</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30001.html">Bytes</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4481,7 +4481,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-w"><span id="30003"></span><a href="../asm/30003.html">30003</a></td>
             <td class="map-length-0">4</td>
             <td class="map-w-desc">
-            <div class="map-entry-title-10">Words</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30003.html">Words</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4492,7 +4492,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-g"><span id="30007"></span><a href="../asm/30007.html">30007</a></td>
             <td class="map-length-0">1</td>
             <td class="map-g-desc">
-            <div class="map-entry-title-10">GSB entry</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30007.html">GSB entry</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4503,7 +4503,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-u"><span id="30008"></span><a href="../asm/30008.html">30008</a></td>
             <td class="map-length-0">1</td>
             <td class="map-u-desc">
-            <div class="map-entry-title-10">Unused</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30008.html">Unused</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4514,7 +4514,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-s"><span id="30009"></span><a href="../asm/30009.html">30009</a></td>
             <td class="map-length-0">9</td>
             <td class="map-s-desc">
-            <div class="map-entry-title-10">Zeroes</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30009.html">Zeroes</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4525,7 +4525,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-t"><span id="30018"></span><a href="../asm/30018.html">30018</a></td>
             <td class="map-length-0">2</td>
             <td class="map-t-desc">
-            <div class="map-entry-title-10">Text</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30018.html">Text</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4557,7 +4557,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="30000"></span><a href="../asm/30000.html">30000</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">Routine</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30000.html">Routine</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4589,7 +4589,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-b"><span id="30001"></span><a href="../asm/30001.html">30001</a></td>
             <td class="map-length-0">2</td>
             <td class="map-b-desc">
-            <div class="map-entry-title-10">Bytes</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30001.html">Bytes</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4600,7 +4600,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-w"><span id="30003"></span><a href="../asm/30003.html">30003</a></td>
             <td class="map-length-0">4</td>
             <td class="map-w-desc">
-            <div class="map-entry-title-10">Words</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30003.html">Words</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4632,7 +4632,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-t"><span id="30018"></span><a href="../asm/30018.html">30018</a></td>
             <td class="map-length-0">2</td>
             <td class="map-t-desc">
-            <div class="map-entry-title-10">Text</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30018.html">Text</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4664,7 +4664,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-u"><span id="30008"></span><a href="../asm/30008.html">30008</a></td>
             <td class="map-length-1">1</td>
             <td class="map-u-desc">
-            <div class="map-entry-title-10">Unused</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30008.html">Unused</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4675,7 +4675,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-s"><span id="30009"></span><a href="../asm/30009.html">30009</a></td>
             <td class="map-length-1">9</td>
             <td class="map-s-desc">
-            <div class="map-entry-title-10">Zeroes</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30009.html">Zeroes</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4718,7 +4718,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="5ba0"></span><a href="../asm/23456.html">23456</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">Routine at 23456</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/23456.html">Routine at 23456</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4754,7 +4754,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="8707"></span><a href="../asm/34567.html">34567</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">Routine at 34567</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/34567.html">Routine at 34567</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4801,7 +4801,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="32768"></span><a href="../asm.html#32768">32768</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">A routine here</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm.html#32768">A routine here</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4812,7 +4812,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-b"><span id="32769"></span><a href="../asm.html#32769">32769</a></td>
             <td class="map-length-0">1</td>
             <td class="map-b-desc">
-            <div class="map-entry-title-10">A data block there</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm.html#32769">A data block there</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4851,10 +4851,10 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <tr>
             <td class="map-page-1">117</td>
             <td class="map-byte-1">48</td>
-            <td class="map-c"><span id="30000"></span><a href="../{}#30000">30000</a></td>
+            <td class="map-c"><span id="30000"></span><a href="../{0}#30000">30000</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">A routine</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../{0}#30000">A routine</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -4931,7 +4931,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="30000"></span><a href="../asm/30000.html">30000</a></td>
             <td class="map-length-1">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-11">Routine</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30000.html">Routine</a></div>
             <div class="map-entry-desc-1">
             <div class="paragraph">
             Return early, return often.
@@ -4945,7 +4945,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-g"><span id="30003"></span><a href="../asm/30003.html">30003</a></td>
             <td class="map-length-1">1</td>
             <td class="map-g-desc">
-            <div class="map-entry-title-11">GSB entry</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30003.html">GSB entry</a></div>
             <div class="map-entry-desc-1">
             </div>
             </td>
@@ -4984,7 +4984,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="45678"></span><a href="../asm/45678.html">B26E</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">Routine at 45678</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/45678.html">Routine at 45678</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5019,7 +5019,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="32768"></span><a href="../asm/32768.html">32768</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">Code</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/32768.html">Code</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5066,7 +5066,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
                 <td class="map-c"><span id="{2}"></span><a href="../asm/{2}.html">{2:05d}</a></td>
                 <td class="map-length-0">1</td>
                 <td class="map-c-desc">
-                <div class="map-entry-title-10"></div>
+                <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/{2}.html"></a></div>
                 <div class="map-entry-desc-0">
                 </div>
                 </td>
@@ -5127,7 +5127,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-w"><span id="30001"></span><a href="../asm/30001.html">30001</a></td>
             <td class="map-length-0">2</td>
             <td class="map-w-desc">
-            <div class="map-entry-title-10">Data</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30001.html">Data</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5138,7 +5138,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-t"><span id="30003"></span><a href="../asm/30003.html">30003</a></td>
             <td class="map-length-0">1</td>
             <td class="map-t-desc">
-            <div class="map-entry-title-10">Message ID</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30003.html">Message ID</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5149,7 +5149,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-t"><span id="30004"></span><a href="../asm/30004.html">30004</a></td>
             <td class="map-length-0">1</td>
             <td class="map-t-desc">
-            <div class="map-entry-title-10">Another message ID</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="../asm/30004.html">Another message ID</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5659,7 +5659,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="65535"></span><a href="65535.html">65535</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">{}</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="65535.html">{}</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5706,7 +5706,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-c"><span id="65535"></span><a href="asm.html#65535">65535</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">{}</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="asm.html#65535">{}</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -5755,10 +5755,10 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <tr>
             <td class="map-page-0">255</td>
             <td class="map-byte-0">255</td>
-            <td class="map-c"><span id="65535"></span><a href="{}#65535">65535</a></td>
+            <td class="map-c"><span id="65535"></span><a href="{0}#65535">65535</a></td>
             <td class="map-length-0">1</td>
             <td class="map-c-desc">
-            <div class="map-entry-title-10">{}</div>
+            <div class="map-entry-title-10"><a class="map-entry-title" href="{0}#65535">{1}</a></div>
             <div class="map-entry-desc-0">
             </div>
             </td>
@@ -6488,7 +6488,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-g"><span id="30000"></span><a href="../asm/30000.html">30000</a></td>
             <td class="map-length-1">1</td>
             <td class="map-g-desc">
-            <div class="map-entry-title-11">GSB entry 1</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30000.html">GSB entry 1</a></div>
             <div class="map-entry-desc-1">
             <div class="paragraph">
             Number of lives.
@@ -6502,7 +6502,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-g"><span id="30001"></span><a href="../asm/30001.html">30001</a></td>
             <td class="map-length-1">2</td>
             <td class="map-g-desc">
-            <div class="map-entry-title-11">GSB entry 2</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30001.html">GSB entry 2</a></div>
             <div class="map-entry-desc-1">
             </div>
             </td>
@@ -6560,7 +6560,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-g"><span id="30000"></span><a href="../asm/30000.html">30000</a></td>
             <td class="map-length-1">1</td>
             <td class="map-g-desc">
-            <div class="map-entry-title-11">GSB entry 1</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30000.html">GSB entry 1</a></div>
             <div class="map-entry-desc-1">
             <div class="paragraph">
             Number of lives.
@@ -6574,7 +6574,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-g"><span id="30001"></span><a href="../asm/30001.html">30001</a></td>
             <td class="map-length-1">2</td>
             <td class="map-g-desc">
-            <div class="map-entry-title-11">GSB entry 2</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30001.html">GSB entry 2</a></div>
             <div class="map-entry-desc-1">
             </div>
             </td>
@@ -6585,7 +6585,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-t"><span id="30003"></span><a href="../asm/30003.html">30003</a></td>
             <td class="map-length-1">1</td>
             <td class="map-t-desc">
-            <div class="map-entry-title-11">Message ID</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30003.html">Message ID</a></div>
             <div class="map-entry-desc-1">
             </div>
             </td>
@@ -6596,7 +6596,7 @@ class HtmlOutputTest(HtmlWriterOutputTestCase):
             <td class="map-t"><span id="30004"></span><a href="../asm/30004.html">30004</a></td>
             <td class="map-length-1">1</td>
             <td class="map-t-desc">
-            <div class="map-entry-title-11">Another message ID</div>
+            <div class="map-entry-title-11"><a class="map-entry-title" href="../asm/30004.html">Another message ID</a></div>
             <div class="map-entry-desc-1">
             </div>
             </td>